split Lwt and non-Lwt code into separate libraries
Introduce xenstore_transport.core and xenstore_transport.lwt.

Code that only wants the Unix backend doesn't need to link Lwt.
This is useful if applications introduce link-time checks to avoid
accidentally linking and using Lwt functions (e.g. such as XAPI, because it has many
blocking functions, and is incompatible with Lwt).

For backwards compatibility xenstore_transport is retained as a library
that links and reexports both core and lwt, so users can move to use the
'.core' library at the own pace.

This effectively undoes the deprecation in, turns out the separate
subpackages would be useful after all:
Fixes: b6d83e1a1 ("Add backwards-compatibility ocamlfind subpackages")

Signed-off-by: Edwin Török <edwin.torok@cloud.com>